refactor: back-button component
